examples/scanner/pyproject.toml
[project]
name = "vantacode-scanner"
version = "0.1.0"
description = "Reference scanner that maps source code to VantaCode taxonomies and galaxies (DKC, OFA, MBT, governance) and emits JSON / CSV / STIX 2.1."
readme = "README.md"
requires-python = ">=3.10"
license = { text = "BSD-2-Clause" }
authors = [{ name = "Cory Kennedy", email = "cory@darkcode.ai" }]
keywords = ["vantacode", "dkc", "ofa", "misp", "stix", "taxonomy", "scanner", "ai-security"]
dependencies = [
    "PyYAML>=6.0.1",
    "click>=8.1.7",
]

[project.optional-dependencies]
stix = ["stix2>=3.0.1"]
all  = ["stix2>=3.0.1"]

[project.scripts]
vantacode-scan = "vantacode_scanner.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["vantacode_scanner"]
